+++ marek_speech_recognition_api/src/recognizer.rs
@@ -3,7 +3,7 @@ use crate::error::SpeechResult;
 pub trait Recognizer {
     fn start(&mut self) -> SpeechResult;
 
-    fn write(&mut self, buffer: &[u8]) -> SpeechResult;
+    fn write(&mut self, buffer: &[i16]) -> SpeechResult;
 
     fn stop(&mut self) -> SpeechResult;
 }
